added ascending ordering to load query repository
