+++ backend/infrastructures/apps/common/models/models.py
@@ -1,7 +1,7 @@
 import time
 
 from django.db import models
-
+from django.utils import timezone
 
 class AutomaticallyTimestampedModel(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
@@ -12,7 +12,7 @@ class AutomaticallyTimestampedModel(models.Model):
 
 
 class ManuallyTimestampedModel(models.Model):
-    created_at = models.DateTimeField(default=time.time(), editable=False)
+    created_at = models.DateTimeField(default=timezone.now(), editable=False)
     updated_at = models.DateTimeField(auto_now=True)
 
     class Meta:
