+++ backend/infrastructures/apps/load/adapters/mappers.py
@@ -10,5 +10,4 @@ def map_data_model_to_output_data_dto(
         full_name=data.data["full_name"],
         age=data.data["age"],
         is_satisfied=data.data["is_satisfied"],
-        #todo add timestamp 
     )
