+++ backend/infrastructures/apps/load/adapters/repositories.py
@@ -18,27 +18,6 @@ from typing import Any
 
 logger = logging.getLogger(__name__)
 
-#TODO move this to extract adapters
-class DjangoDataDomainRepository(domain_repositories.AbstractDataDomainRepository):
-    """
-    See description of parent class to get more details.
-    """
-
-    def create(self, data: list[load_queries.OutputData]) -> None:
-        Data.objects.bulk_create(
-            [
-                Data(
-                    data={
-                        "full_name": output_data.full_name,
-                        "age": output_data.age,
-                        "is_satisfied": output_data.is_satisfied,
-                    }
-                )
-                for output_data in data
-            ]
-        )
-
-
 class DjangoDataQueryRepository(query_ports.AbstractDataQueryRepository):
     """
     See description of parent class to get more details.
@@ -66,16 +45,16 @@ def _get_django_output_data_filters(filters: query_ports.OutputDataFilters) -> d
     django_filters: dict[str, Any] = {}
 
     if filters.age is not None:
-        django_filters["age"] = filters.age
+        django_filters["data__age"] = filters.age
 
     if filters.is_satisfied is not None:
-        django_filters["is_satisfied"] = filters.is_satisfied
+        django_filters["data__is_satisfied"] = filters.is_satisfied
 
     if filters.timestamp_from is not None:
-        django_filters["timestamp_from__gte"] = filters.timestamp_from
+        django_filters["created_at__gte"] = filters.timestamp_from
 
     if filters.timestamp_to is not None:
-        django_filters["timestamp_to__lte"] = filters.timestamp_to
+        django_filters["created_at__lte"] = filters.timestamp_to
 
     return django_filters
 
@@ -84,6 +63,6 @@ def _get_django_output_data_ordering(ordering: query_ports.OutputDataOrdering) -
     django_ordering: dict[str, ordering_dtos.Ordering] = {}
 
     if ordering.timestamp is not None:
-        django_ordering["timestamp"] = ordering.timestamp
+        django_ordering["created_at"] = ordering.timestamp
 
     return common_ordering.get_django_ordering(django_ordering)
