+++ backend/infrastructures/apps/load/views/serializers.py
@@ -9,5 +9,3 @@ class OutputDataReadSerializer(serializers.Serializer):
     full_name = serializers.CharField()
     age = serializers.IntegerField()
     is_satisfied = serializers.BooleanField()
-    created_at = serializers.DateField()
-    updated_at = serializers.DateField()
