+++ backend/infrastructures/apps/load/views/urls.py
@@ -3,6 +3,7 @@ from rest_framework.routers import DefaultRouter
 
 from .views import LoadViewSet
 
+
 load_router = DefaultRouter()
 load_router.register(prefix=r"", viewset=LoadViewSet, basename="load")
