+++ backend/infrastructures/config/settings/injections_configurations.py
@@ -1,7 +1,7 @@
 import inject
 
 from ...apps.load.adapters import repositories as query_load_repositories
-from ...apps.load.adapters import units_of_work as load_unit_of_work
+from ...apps.extract.adapters import units_of_work as load_unit_of_work
 
 
 def inject_config(binder: inject.Binder):
